add random apply transform base class
