+++ torchvision/prototype/transforms/_geometry.py
@@ -12,21 +12,11 @@ from torchvision.transforms.functional import pil_to_tensor
 from torchvision.transforms.transforms import _setup_size, _interpolation_modes_from_int
 from typing_extensions import Literal
 
+from ._transform import _RandomApplyTransform
 from ._utils import query_image, get_image_dimensions, has_any, is_simple_tensor
 
 
-class RandomHorizontalFlip(Transform):
-    def __init__(self, p: float = 0.5) -> None:
-        super().__init__()
-        self.p = p
-
-    def forward(self, *inputs: Any) -> Any:
-        sample = inputs if len(inputs) > 1 else inputs[0]
-        if torch.rand(1) >= self.p:
-            return sample
-
-        return super().forward(sample)
-
+class RandomHorizontalFlip(_RandomApplyTransform):
     def _transform(self, input: Any, params: Dict[str, Any]) -> Any:
         if isinstance(input, features.Image):
             output = F.horizontal_flip_image_tensor(input)
@@ -45,18 +35,7 @@ class RandomHorizontalFlip(Transform):
             return input
 
 
-class RandomVerticalFlip(Transform):
-    def __init__(self, p: float = 0.5) -> None:
-        super().__init__()
-        self.p = p
-
-    def forward(self, *inputs: Any) -> Any:
-        sample = inputs if len(inputs) > 1 else inputs[0]
-        if torch.rand(1) > self.p:
-            return sample
-
-        return super().forward(sample)
-
+class RandomVerticalFlip(_RandomApplyTransform):
     def _transform(self, input: Any, params: Dict[str, Any]) -> Any:
         if isinstance(input, features.Image):
             output = F.vertical_flip_image_tensor(input)
@@ -371,11 +350,11 @@ class Pad(Transform):
             return input
 
 
-class RandomZoomOut(Transform):
+class RandomZoomOut(_RandomApplyTransform):
     def __init__(
         self, fill: Union[float, Sequence[float]] = 0.0, side_range: Tuple[float, float] = (1.0, 4.0), p: float = 0.5
     ) -> None:
-        super().__init__()
+        super().__init__(p=p)
 
         if fill is None:
             fill = 0.0
@@ -384,8 +363,6 @@ class RandomZoomOut(Transform):
         self.side_range = side_range
         if side_range[0] < 1.0 or side_range[0] > side_range[1]:
             raise ValueError(f"Invalid canvas side range provided {side_range}.")
-
-        self.p = p
 
     def _get_params(self, sample: Any) -> Dict[str, Any]:
         image = query_image(sample)
@@ -411,10 +388,3 @@ class RandomZoomOut(Transform):
     def _transform(self, input: Any, params: Dict[str, Any]) -> Any:
         transform = Pad(**params, padding_mode="constant")
         return transform(input)
-
-    def forward(self, *inputs: Any) -> Any:
-        sample = inputs if len(inputs) > 1 else inputs[0]
-        if torch.rand(1) >= self.p:
-            return sample
-
-        return super().forward(sample)
