+++ torchvision/prototype/transforms/_transform.py
@@ -2,6 +2,7 @@ import enum
 import functools
 from typing import Any, Dict
 
+import torch
 from torch import nn
 from torchvision.prototype.utils._internal import apply_recursively
 from torchvision.utils import _log_api_usage_once
@@ -34,3 +35,20 @@ class Transform(nn.Module):
             extra.append(f"{name}={value}")
 
         return ", ".join(extra)
+
+
+class _RandomApplyTransform(Transform):
+    def __init__(self, *, p: float = 0.5) -> None:
+        if not (0.0 <= p <= 1.0):
+            raise ValueError("`p` should be a floating point value in the interval [0.0, 1.0].")
+
+        super().__init__()
+        self.p = p
+
+    def forward(self, *inputs: Any) -> Any:
+        sample = inputs if len(inputs) > 1 else inputs[0]
+
+        if torch.rand(1) >= self.p:
+            return sample
+
+        return super().forward(sample)
